+++ LuiExtended/modules/UnitFrames/_StaticControls/Group.lua
@@ -8,10 +8,14 @@ local FrameObject = LUIE_CustomFrameObject
 function FrameObject:UpdateGroupFrameStaticControls()
     FrameObject.UpdateStaticControlRoleIcon(self)
     FrameObject.UpdateStaticControlClassIcon(self)
-    self.friendIcon:SetHidden(true)
-    FrameObject.UpdateTopInfoOverlandIcon(self)
-    FrameObject.UpdateStaticControlNameLabel(self, "group")
-    FrameObject.UpdateStaticControlLevelRow(self)
-    FrameObject.LayoutTopInfoSmallGroup(self)
+    if self.friendIcon then
+        self.friendIcon:SetHidden(true)
+    end
+    if FrameObject.HasCustomTopInfoFrameCategory(self) then
+        FrameObject.UpdateTopInfoOverlandIcon(self)
+        FrameObject.UpdateStaticControlNameLabel(self, "group")
+        FrameObject.UpdateStaticControlLevelRow(self)
+        FrameObject.LayoutTopInfoSmallGroup(self)
+    end
     FrameObject.UpdateStaticControlDeadAndGroupAlpha(self)
 end